updated data type decimal to float

+++ server/db/migrations/20200507054824_income.js
@@ -7,7 +7,7 @@ exports.up = knex =>
     t.string('income_name')
     t.string('frequency').defaultTo('Weekly')
     t.string('income_type')
-    t.decimal('income_amount')
+    t.float('income_amount')
     t.boolean('active').defaultTo(true)
   })
